% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pomodoror.R
\name{pomodoror}
\alias{pomodoror}
\title{Pomodoror}
\usage{
pomodoror(pomodoros = 4, work_length = 20, break_length = 10,
  work_start = "mario", work_end = "fanfare")
}
\arguments{
\item{pomodoros}{Integer value for the number of repetitions.}

\item{work_length}{Amount of time in seconds to work before a break.}

\item{break_length}{Amount of time in seconds to break before getting back to work.}

\item{work_start}{Noise to make to signal beginning of work time.}

\item{work_end}{Noise to make to signal end of work time/beginning of break time.}
}
\description{
Write more productively with Pomodoror!
}
